% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastrls-package.R
\docType{package}
\name{fastrls-package}
\alias{fastrls-package}
\title{fastrls: Fast Recursive Directory Listing}
\description{
Fast recursive directory listing using Rust crates walkdir \url{https://github.com/BurntSushi/walkdir} and jwalk \url{https://github.com/Byron/jwalk}.
}
\author{
\strong{Maintainer}: Ben Raymond \email{ben.raymond@aad.gov.au}

}
\keyword{internal}
