% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkers.R
\name{fastrls}
\alias{fastrls}
\title{Recursively list files and directories}
\usage{
fastrls(dir, include_dirs = TRUE, follow_links = TRUE, n_threads = 2L)
}
\arguments{
\item{dir}{string: starting directory}

\item{include_dirs}{logical: if \code{TRUE}, include directories in the results}

\item{follow_links}{logical if \code{TRUE}, follow symbolic links}

\item{n_threads}{integer: if > 1, use multiple threads. Might be counter-productive on hardware that can't cope with simultaneous requests (spinning disks?)}
}
\value{
A character vector
}
\description{
Recursively list files and directories
}
